+++ CHANGELOG.md
@@ -245,5 +245,6 @@
 * Fixed NRE when reloading save on SpriteActorAnimationComponent
 * Added HasVideo API to VideoModule
 * Added ConditionalFilter that filters action special triggers by evaluating a Conditional
+* Added explicit EndQuest API (SetQuestStage no longer sends "quest finished" message on negative quest stages)
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [Untested]
 * Fixed dialogue reactivating hidden objects even if they were initially inactive [Untested]